.profile.d/README.functions - 移除 Posix shell 沒有規範的 readonly function 宣告法
```
Read-only functions were considered, but they were omitted as not being historical practice or particularly useful. Furthermore, functions must not be read-only across invocations to preclude "spoofing" (spoofing is the term for the practice of creating a program that acts like a well-known utility with the intent of subverting the real intent of the user) of administrative or security-relevant (or security-conscious) shell scripts.
```

Reference: Shell Command Language <http://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#readonly>

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>